new_backend/requirements.txt
alabaster==0.7.12
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==4.2.1
bleach==4.1.0
blinker==1.9.0
brial==1.0.2
Brotli==1.1.0
cachelib==0.13.0
celery==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.11.0
cypari2==2.1.2
Cython==0.29.28
datasketch==1.6.5
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fasttext==0.9.3
filelock==3.6.0
Flask==3.0.3
Flask-Compress==1.17
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.29.1
fpylll==0.5.2.dev0
fs==2.4.12
fsspec==2024.10.0
gast==0.5.2
gmpy2==2.1.2
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.5.3
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.3
memory-allocator==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
msgspec==0.19.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
notebook==6.4.8
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.88.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgconfig==1.5.5
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
pplpy==0.8.7
primecountpy==0.1.0
prometheus-client==0.9.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
py==1.10.0
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybind11==2.13.6
pycurl==7.44.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMySQL==1.1.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythran==0.10.0
pytz==2024.2
PyYAML==5.4.1
pyzmq==22.3.0
redis==5.2.1
regex==2024.9.11
requests==2.25.1
roman==3.3
rpy2==3.4.5
safetensors==0.4.5
sagemath-standard==9.5
sagenb-export==3.2
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentence-transformers==3.2.1
service-identity==18.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
SQLAlchemy==2.0.41
supervisor==4.2.1
sympy==1.13.1
systemd-python==234
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tornado==6.1
tox==3.21.4
tqdm==4.66.6
traitlets==5.1.1
transformers==4.46.1
triton==3.1.0
Twisted==22.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==2.1
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
vine==5.1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==2.0.0
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.23.0
